requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
html5lib>=1.1
pytz>=2023.3
brotli>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
apscheduler>=3.10.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

